+++ fuc/api/pyfq.py
@@ -1,7 +1,7 @@
 """
 The pyfq submodule is designed for working with FASTQ files (both zipped
-and unzipped). It implements ``pyfq.FqFrame`` which stores FASTQ data as
-``pandas.DataFrame`` to allow fast computation and easy manipulation.
+and unzipped). It implements the ``pyfq.FqFrame`` class which stores FASTQ
+data as ``pandas.DataFrame`` to allow fast computation and easy manipulation.
 """
 
 import gzip